chore: add leaf serde
