+++ crates/committer/src/patricia_merkle_tree/node_data.rs
@@ -1,2 +1,3 @@
 pub mod inner_node;
 pub mod leaf;
+pub mod leaf_serde;
